requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
configobj==5.0.9
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
httpx-ws==0.7.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-ollama==0.3.6
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
lmstudio==1.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
msgspec==0.19.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.1.2
numpy==2.3.2
ollama==0.5.3
openai==1.100.2
orjson==3.11.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.0
selenium==4.35.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
streamlit==1.48.1
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
zstandard==0.24.0
